+++ Tests/LegacyUnit/FrontEnd/AddToFavoritesButtonViewTest.php
@@ -27,7 +27,7 @@ class tx_realty_FrontEnd_AddToFavoritesButtonViewTest extends \Tx_Phpunit_TestCa
         /** @var TypoScriptFrontendController $frontEndController */
         $frontEndController = $GLOBALS['TSFE'];
         $this->fixture = new tx_realty_pi1_AddToFavoritesButtonView(
-            ['templateFile' => 'EXT:realty/pi1/tx_realty_pi1.tpl.htm'],
+            ['templateFile' => 'EXT:realty/Resources/Private/Templates/FrontEnd/Plugin.html'],
             $frontEndController->cObj
         );
     }