+++ Tests/LegacyUnit/FrontEnd/DefaultListViewTest.php
@@ -66,7 +66,7 @@ class tx_realty_FrontEnd_DefaultListViewTest extends \Tx_Phpunit_TestCase
 
         $this->fixture = new tx_realty_pi1_DefaultListView(
             [
-                'templateFile' => 'EXT:realty/pi1/tx_realty_pi1.tpl.htm',
+                'templateFile' => 'EXT:realty/Resources/Private/Templates/FrontEnd/Plugin.html',
                 'pages' => $this->systemFolderPid,
                 'showGoogleMaps' => 0,
             ],