+++ Tests/LegacyUnit/FrontEnd/FilterFormTest.php
@@ -27,7 +27,7 @@ class tx_realty_FrontEnd_FilterFormTest extends \Tx_Phpunit_TestCase
 
         $this->fixture = new tx_realty_filterForm(
             [
-                'templateFile' => 'EXT:realty/pi1/tx_realty_pi1.tpl.htm',
+                'templateFile' => 'EXT:realty/Resources/Private/Templates/FrontEnd/Plugin.html',
                 'displayedSearchWidgetFields' => 'site',
             ],
             $this->getFrontEndController()->cObj