+++ Tests/LegacyUnit/FrontEnd/GoogleMapsViewTest.php
@@ -92,7 +92,7 @@ class tx_realty_FrontEnd_GoogleMapsViewTest extends \Tx_Phpunit_TestCase
 
         $this->fixture = new tx_realty_pi1_GoogleMapsView(
             [
-                'templateFile' => 'EXT:realty/pi1/tx_realty_pi1.tpl.htm',
+                'templateFile' => 'EXT:realty/Resources/Private/Templates/FrontEnd/Plugin.html',
                 'defaultCountryUID' => 54,
             ],
             $this->getFrontEndController()->cObj,