+++ Tests/LegacyUnit/FrontEnd/ObjectsByOwnerListViewTest.php
@@ -55,7 +55,7 @@ class tx_realty_FrontEnd_ObjectsByOwnerListViewTest extends \Tx_Phpunit_TestCase
         $frontEndController = $GLOBALS['TSFE'];
         $this->fixture = new tx_realty_pi1_ObjectsByOwnerListView(
             [
-                'templateFile' => 'EXT:realty/pi1/tx_realty_pi1.tpl.htm',
+                'templateFile' => 'EXT:realty/Resources/Private/Templates/FrontEnd/Plugin.html',
                 'pages' => $this->systemFolderPid,
             ],
             $frontEndController->cObj,