+++ Tests/LegacyUnit/FrontEnd/OffererListTest.php
@@ -50,7 +50,7 @@ class tx_realty_FrontEnd_OffererListTest extends \Tx_Phpunit_TestCase
         $frontEndController = $GLOBALS['TSFE'];
         $this->fixture = new tx_realty_offererList(
             [
-                'templateFile' => 'EXT:realty/pi1/tx_realty_pi1.tpl.htm',
+                'templateFile' => 'EXT:realty/Resources/Private/Templates/FrontEnd/Plugin.html',
                 'what_to_display' => 'offerer_list',
                 'userGroupsForOffererList' => $this->feUserGroupUid,
                 'displayedContactInformation' => 'usergroup,offerer_label',