+++ Tests/LegacyUnit/FrontEnd/OffererViewTest.php
@@ -29,7 +29,7 @@ class tx_realty_FrontEnd_OffererViewTest extends \Tx_Phpunit_TestCase
         /** @var TypoScriptFrontendController $frontEndController */
         $frontEndController = $GLOBALS['TSFE'];
         $this->fixture = new tx_realty_pi1_OffererView(
-            ['templateFile' => 'EXT:realty/pi1/tx_realty_pi1.tpl.htm'],
+            ['templateFile' => 'EXT:realty/Resources/Private/Templates/FrontEnd/Plugin.html'],
             $frontEndController->cObj
         );
         $this->fixture->setConfigurationValue(